backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0
scikit-learn==1.3.2
numpy==1.24.3
email-validator==2.1.0
